+++ hosts/eris/hardware-configuration.nix
@@ -21,5 +21,4 @@
   };
 
   swapDevices = [{ device = "/dev/sdb"; }];
-
 }