+++ hosts/sigma/home.nix
@@ -1,11 +1,6 @@
 { ... }:
 
 {
-  home.file.".ssh/config".text = ''
-    Host *
-      IdentityAgent ~/.1password/agent.sock
-  '';
-
   xsession.windowManager.i3.config.workspaceOutputAssign = [{
     workspace = "101";
     output = "eDP-1";