+++ hosts/upsilon/home.nix
@@ -1,11 +1,6 @@
 { keys, ... }:
 
 {
-  home.file.".ssh/config".text = ''
-    Host *
-      IdentityAgent "~/Library/Group Containers/2BUA8C4S2C.com.1password/t/agent.sock"
-  '';
-
   programs.git.extraConfig = {
     user.signingKey = keys.users."michael.hoang_upsilon";
     gpg.format = "ssh";