+++ modules/graphical.nix
@@ -33,5 +33,21 @@
     programs._1password.enable = true;
   };
 
-  hmModule = { pkgs, ... }: { programs.vscode.package = pkgs.vscode; };
+  hmModule = { pkgs, ... }: {
+    programs.vscode.package = pkgs.vscode;
+
+    home.file.".ssh/config".text = let
+      _1password-agent = if pkgs.stdenv.hostPlatform.isDarwin then
+        "~/Library/Group Containers/2BUA8C4S2C.com.1password/t/agent.sock"
+      else
+        "~/.1password/agent.sock";
+    in ''
+      Match host * exec "test -z $SSH_TTY"
+        IdentityAgent "${_1password-agent}"
+        ForwardAgent yes
+
+      Host *
+        ServerAliveInterval 120
+    '';
+  };
 }